InputView 싱글톤 구현
