+++ src/main/java/Controller/ModelHandler.java
@@ -21,4 +21,6 @@ public class ModelHandler {
     public static ModelHandler getInstance() {
         return Singleton.INSTANCE;
     }
+
+
 }
