--- src/main/java/View/inputView.java
@@ -1,20 +1,0 @@
-package View;
-
-import Controller.ModelHandler;
-import camp.nextstep.edu.missionutils.Console;
-
-public class inputView {
-    private ModelHandler moder = ModelHandler.getInstance();
-
-    public void inputPrice() {
-
-    }
-
-    public void inputWinningNumber() {
-
-    }
-
-    public void inputBonusNumber() {
-
-    }
-}
